requirements.txt
# FastAPI & Web
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2

# Auth & Security 
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.5.2

# LangChain & RAG
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.0
langchain-huggingface==0.1.0
langchain-google-genai==2.0.0
google-generativeai==0.7.2
langchain-ollama
# Vector Stores
qdrant-client==1.11.0

# Embeddings & LLMs
sentence-transformers==3.1.0
transformers==4.45.0
torch==2.4.0
accelerate==0.34.0
openai==1.51.0
tiktoken==0.8.0

# PDF Processing
pypdf==4.3.1
pdfplumber==0.11.4

# Utilities
tenacity==8.5.0
email-validator==2.2.0
asyncpg==0.29.0
bcrypt==4.2.0